+++ frontend/src/ts/commandline/lists/result-saving.ts
@@ -1,0 +1,38 @@
+import * as TestState from "../../test/test-state";
+import * as ModesNotice from "../../elements/modes-notice";
+
+const subgroup: MonkeyTypes.CommandsSubgroup = {
+  title: "Result saving...",
+  list: [
+    {
+      id: "setResultSavingOff",
+      display: "off",
+      alias: "disabled",
+      exec: (): void => {
+        TestState.setSaving(false);
+        ModesNotice.update();
+      },
+    },
+    {
+      id: "setResultSavingOn",
+      display: "on",
+      alias: "enabled",
+      exec: (): void => {
+        TestState.setSaving(true);
+        ModesNotice.update();
+      },
+    },
+  ],
+};
+
+const commands: MonkeyTypes.Command[] = [
+  {
+    id: "setResultSaving",
+    display: "Result saving...",
+    icon: "fa-save",
+    alias: "results",
+    subgroup,
+  },
+];
+
+export default commands;
